requirements.txt
# --- 核心框架 ---
langgraph               # 必选：多智能体状态机编排
langchain-openai        # 必选：用于调用智谱 AI 的 OpenAI 兼容接口
python-dotenv           # 必选：管理环境变量

# --- 数据处理 ---
pydantic>=2.0           # 必选：数据建模和格式校验
httpx                   # 必选：高性能异步网络请求
feedparser              # 必选：解析 RSS 订阅源

# --- 存储 ---
chromadb                # 必选：本地轻量级向量数据库（记忆系统）